:recycle: viewCaretを分離した
